unit testing, code cleanup, improved sample flag iplementation

+++ src/wav/mod.rs
@@ -9,7 +9,7 @@ const HEADER_SIZE: u8 = 44;
 
 pub fn build_header(
     smp_rate: u32,  // sample rate
-    smp_bits: u16,  // bits per sample
+    smp_bits: u8,  // bits per sample
     pcm_len: u32,   // length of byte array
 ) -> [u8; HEADER_SIZE as usize]
 {
@@ -18,9 +18,9 @@ pub fn build_header(
     let wav_type:           u16 = 1;                        // 1 = pcm
     let wav_flag_ms:        u16 = 0x01;                     // mono/stereo 0x01 = mono, 0x02 = stereo
     let sample_frequency:   u32 = smp_rate;
-    let bytes_sec:          u32 = smp_rate * wav_flag_ms;   // sample_rate * channels (DOUBLE CHECK)
+    let bytes_sec:          u32 = smp_rate * wav_flag_ms as u32;   // sample_rate * channels (DOUBLE CHECK)
     let block_align:        u16 = 0x01;                     // can be anything really
-    let bits_sample:        u16 = smp_bits;
+    let bits_sample:        u16 = smp_bits as u16;
     let file_size:          u32 = HEADER_SIZE as u32 + (pcm_len * (bits_sample / 8) as u32) - 8;
     let size_of_chunk:      u32 = pcm_len * (bits_sample / 8) as u32;
